refactor: move formatDate function to utils
